Add support for persistent disk when booting from ISO image
When booting from a LiveCD image there is no point in generating
a diffDisk because the OS is running from RAM, not from the disk.

In that case just add diffDisk as an independent data volume and
move directories that should be persisted on there. Currently:
/home, /usr/local, /etc/containerd, and /var/lib/containerd.

This data-volume logic is currently only enabled for Alpine because
it hasn't been tested anywhere else yet.

Additional, if the configured disk size is 0, no diffDisk will be
allocated and changes will be written back to the baseDisk (or
discarded, if the baseDisk is an ISO image).

Signed-off-by: Jan Dubois <jan.dubois@suse.com>