Extract minimal set of lima and qemu files to install on a clean machine
Installing qemu via homebrew installs 846MB if stuff under /usr/local/Cellar,
571MB of which are the qemu cellar alone.

This is clearly excessive for bundling lime with an app that simply wants
to deploy VMs.

The attached script uses dtrace to observe which files are actually opened
by lima and qemu, and then collects just those. It collects the corresponding
symlinks and also modifies all Mach-O binaries to reference dylibs relative
to the path of the executable, so the files can be relocated anywhere.

There are still hard-coded paths to things under /user/local inside the binaries,
but testing shows that lima works with the bundled files as-is.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>