Resolve /usr/local/bin symlinks in hacks/lima-and-qemu
All dylib references in the binaries and libs are updated to use
"@executable_path/.." to locate libs. In macOS Monterey (Public Beta)
this is resolved not relative to the symlink, but relative to the
symlink target.

To create a distro that works on Catalina/Big Sur *and* Monterey, the
symlinks in the bin directory are resolved before packaging.

Symlinks still need to be stored for at least /usr/local/share/qemu so
that qemu will still be able to find its data files.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>