Provide explicit DNS override
In some situations (e.g. VPN) the user may need to explicitly
set the DNS addresses; even the fallback to the "en0" settings
does not work.

This cannot be done in a provisioning script because they run
after the boot scripts, but working DNS is required to install
additional dependencies.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>