Add /opt/vde/bin/vde_vmnet and prerequites to lima-and-qemu tarball
It is kind-of heavy-handed by hardcoding the 2 dependencies, but
extending the opensnoop processing just to deal with /opt/vde would
be rather convoluted (it would only cover the dylib; the executables
have to be added manually anyways).

Installation will have to move the `vde` directory from the tarball
into a location only root can write to.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>