Replace localhost in proxy settings with the gateway address
Lima uses process environment settings to override system proxy
settings, and settings from the `env` key in lima.yaml.

This does not work when the proxy is running on a localhost address,
so this commit replaces localhost with the slirp gateway address.

The new `useProcessProxies` setting can also be used to disable using
the process environment variables to override proxy settings completely.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>