Add IdentityFile, InstanceDir, and LimaHome fields to limactl list format.
Update examples/docker.yaml to show how this can be used to
construct the instance-specific path of forwarded sockets.

Also adds InstanceDir and Name fields to the portForwards[*].hostSocket
field (for symmetry with `), and deprecates the LimaHome and Instance
fields (but keeps supporting them).

The fact that relative paths in hostSocket will be prefixed with
{{.InstanceDir}}/sock is no longer documented but still supported.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>