Allow Proxy ports to be marshalled as strings
It is not clear how/why this can happen, but we have seen at least 2 reports
from users where a missing port was encoded as an empty string, and not as
the number 0. It is unclear if on those systems non-0 ports would be encoded
as strings or numbers, so this commit will handle both.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>