Sync the containerd files for Kubernetes
Upstream includes configuration for CRI and CNI,
so make sure that is added to the nerdctl install.
Currently using "flannel" for multi-node and VXLAN,
but template can be edited to use "containerd-net".

The image endpoint defaults to the runtime endpoint,
so use the same configuration file as upstream has.
Currently we use kubernetes-cni which is a slightly
older version so don't bump the cni version (yet).

Signed-off-by: Anders F Björklund <anders.f.bjorklund@gmail.com>