Treat AAAA queries as A queries when IPv6 is disabled
A "correct" answer would be to set `handled = true` and return a NODATA response.
Unfortunately some older resolvers use a slow random source to set the transaction id.
This creates a problem on M1 computers, which are too fast for that implementation:
Both the A and AAAA queries might end up with the same id. Returning NODATA for AAAA
is faster, so would arrive first, and be treated as the response to the A query.
To avoid this, we will treat an AAAA query as an A query when IPv6 has been disabled.
This way it is either a valid response for an A query, or the A records will be discarded
by a genuine AAAA query, resulting in the desired NODATA response.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>