Add LIMA_CIDATA_HOME_MOUNTPOINT
This can be used by provisioning scripts to locate the mount point of
the users home directory. We can't just use `/Users/$LIMA_CIDATA_USER`
because we don't know if the host is Linux or macOS, so it could be
`/home/$LIMA_CIDATA_USER` or something completely custom. Also
`$LIMA_CIDATA_USER` is the username inside the guest and might be set to
`lima` if the username on the macOS host is not a valid Linux username.

Example usage to configure client certs for the docker daemon:

  ln -s $LIMA_CIDATA_HOME_MOUNTPOINT/.docker/certs.d /etc/docker/certs.d

Signed-off-by: Jan Dubois <jan.dubois@suse.com>