limactl: remove `limactl inspect` (not a breaking change)
The `limactl inspect` command was added in commit 75594fa33bd73a2cc80a389993c7573b4bf3ec58
(PR 1249), but this is no longer useful as initially intended, because
`limactl inspect --format=json` and `limactl ls --format=json` produces
the exactly same output.

The `limactl inspect` command was only present in the master branch,
so this removal is not a breaking change.

Fix issue 1282

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>