Option to upgrade packages and reboot if required
This will run "apt upgrade" or "dnf upgrade" on the first boot,
to make sure that all packages are updated to the latest version.

It adds to the initial startup time and to the size of the diffdisk,
so it is better to update the basedisk with a newer cloud release.

Signed-off-by: Anders F Björklund <anders.f.bjorklund@gmail.com>