hostagent: Handle SIGTERM signal for better compatibility with macOS launchd
This commit modifies the hostagent to handle not only the SIGINT (Interrupt) signal, but also the SIGTERM signal. This allows the hostagent to gracefully shut down when it receives a SIGTERM signal, which is a common way to request a process to terminate. This change improves the compatibility with macOS launchd, which uses SIGTERM to stop services.

The following are the steps to set up the Docker instance of Lima to start at login using launchd on macOS:
```bash
# Create the docker instance
limactl start template://docker --vm-type vz --rosetta --network vzNAT --tty=false
# Stop the docker instance
limactl stop docker
# Create the launchd plist
echo '<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
<plist version="1.0">
<dict>
	<key>Label</key>
	<string>io.lima-vm.docker</string>
	<key>ProgramArguments</key>
	<array>
		<string>'"$(which limactl)"'</string>
		<string>hostagent</string>
		<string>--pidfile</string>
		<string>ha.pid</string>
		<string>--socket</string>
		<string>ha.sock</string>
		<string>docker</string>
	</array>
	<key>RunAtLoad</key>
	<true/>
	<key>StandardErrorPath</key>
	<string>ha.stderr.log</string>
	<key>StandardOutPath</key>
	<string>ha.stdout.log</string>
	<key>WorkingDirectory</key>
	<string>'"$(limactl list docker --format '{{.Dir}}')"'</string>
</dict>
</plist>' > ~/Library/LaunchAgents/io.lima-vm.docker.plist
# Load the launchd plist and launchd will start the docker instance
launchctl load ~/Library/LaunchAgents/io.lima-vm.docker.plist
```

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>