Reinstall apk packages on reboot
Because we are running from a RAM disk, any packages not installed
from the ISO are removed after a reboot. They are still registered
in /etc/apk/world (which is on /mnt/data/etc/apk/world), and the
packages themselves are cached on /mnt/data/apk/cache, so reinstalling
is quick and easy.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>