Never retain ownership during installation
By default, tar will retain ownership when run as root. Even if built
as a user, the install phase should not install binaries owned by that
user. To suppress this, we pass `--no-same-owner` to tar.

While at it, I made it use the `TAR` environment variable during
installation as well.

Signed-off-by: Dan Villiom Podlaski Christiansen <danchr@gmail.com>