limayaml: add a `param` field for defining variables used to customize scripts and other elements within `lima.yaml`.
The defined variables can be referenced within `lima.yaml` as `{{.Param.Key}}`.
The fields where this can be utilized are as follows:
- `provision[%d].script`
- `probes[%d].script`
- `copyToHost[%d].{guest,host}`
- `portForwards[%d].{guestSocket,hostSocket}`

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>

limayaml: add a check to verify whether the variables defined in `param` are actually used

This check needs to be performed before the template is expanded, so it should be added before calling `FillDefault()`.

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>

limayaml: add `{{if .Param.rootful}}{{else}}{{end}}` pattern to `TestValidateParamIsUsed`

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>

default.yaml: add a comment on the `probes` that supports variable substitution

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>