Show progress only when logging to terminal
Logging progress to file is not useful and it also breaks the log
format, adding empty line before the first progress, and leaving the
last progress at the start of the next log line.

Fix the terminal check to use os.Stderr, since both pb and logrus use
os.Stderr. When not using a terminal, hide the progress bar.

Fixes #2581
Signed-off-by: Nir Soffer <nsoffer@redhat.com>