limayaml/defaults_test.go: fix incorrect `expect.Mounts[0].Virtiofs.QueueSize`
`expect.Mounts[0]` was the same instance as each of `y.Mounts[0]` and `d.Mounts[0]`, so the test passed regardless of the value set for `Virtiofs.QueueSize`.
The default for `Virtiofs.QueueSize` is `nil`, but I'll keep the code that explicitly sets it to `nil` to make it clear that `nil` is expected.

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>