Suppress error if file is already deleted
We have a race (time of check, time of use) when deleting a vm with
--force. We first list the file in the instance dir, and then delete
matching files. "ssh.sock" is consistently already deleted when we try
to remove it, creating unwanted noise in the log.

"ssh.sock" is probably removed by other code that may need need to
remove it. To make this easy to improve we still log the event in debug
level.

Fixes: #2635
Signed-off-by: Nir Soffer <nsoffer@redhat.com>