Don't call tail.Cleanup()
It is only supposed to be called from a process exit handler,
and exists as a workaround to a Linux kernel bug that _may_ have
existed in some kernels back in early 2014.

Calling tail.Cleanup() seems to prevent the process from receiving
any further fsnotify notification for the same file again (at least
on macOS).

Signed-off-by: Jan Dubois <jan.dubois@suse.com>