CI: temporarily remove ArchLinux (and demote it from Tier 1)
Workaround for issue 2714

The "user session is ready for ssh" phase can never finishes with the cloud-init
24.3.1-1 package due to a regression that breaks execution of
`/var/lib/cloud/scripts/per-boot/` scripts.

This commit also demotes ArchLinux from Tier 1;
it is practically impossible to maintain good stability as the packages are too
frequently updated.

After the regression is fixed we may bring ArchLinux to Tier 2, but
probably never to Tier 1 again.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>