20-rootless-base.sh: Fix ownership of the user home directory when created by root.
When mount points exist in the user's home directory, both the home directory
and the mount points are created by root before the user is created.
This results in the home directory being owned by root.

This change ensures that the ownership of the home directory and its contents
(on the same filesystem) is updated to the correct user.

If changing the owner of any mount points fails, the error is ignored.

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>