Refactor limayaml.Save and store.SaveYAML → limayaml.Marshal
`store.SaveYAML` was the only caller of `limayaml.Save`, and was only called
from `main.validateAction`.

The name `Save` implies writing to a file, so `Marshal` seems like a more
logical choice for the name. And since it is marshalling `limayaml` structs
it should be in the `limayaml` package and not in `store`.

Also move `limayaml.unmarshalYAML` from `load.go` to `marshal.go` for consistency.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>