Fix windows WithDirLock to exclusive lock
We use LOCKFILE_FAIL_IMMEDIATELY, creating a shared lock that fail
immediately if the lock cannot be acquired. This does not match the
behavior of the unix version and does not make sense.

Using now LOCKFILE_EXCLUSIVE_LOCK to acquire exclusive lock blocking
until the lock can be acquired.

Add constants for the flags and add comments for the argument names to
prevent future errors.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>