networks.Validate() requires that socket_vmnet is owned by root
This restriction was weakened by https://github.com/lima-vm/lima/pull/1220
to only require the file and directories to be owned by the admin user, but
that configuration is not secure.

If users are willing to run an insecure configuration, then they can always
enable password-less sudo, which does not need a sudoers file at all.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>