Add failing redownload test
We tested that modified last modified time on the server cause a
redownload, but we did not test that after a redownload we update the
cache, so the next attempt will used the cached download.

Add a failing test verifying the issue, and improve test comments and
configuration to make it more clear.

Part-of: #2902
Signed-off-by: Nir Soffer <nsoffer@redhat.com>