Un-deprecate non-strict YAML
We still warn the user because it could be due to typos, but we no longer
threaten to make strict mode the default.

Allowing non-strict YAML is very useful for maintainers switching between
feature branches. Failing on non-strict YAML would make it hard to e.g.
delete instances that have been created by a different branch that introduces
a new field.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>