Disable GA serial port in QEMU as it is not used by GA
GA was switched back to use SSH socket forwarding due to stability issues,
but QEMU still exposes the serial as a Unix socket, which is

1) unused
2) removed from FS and replaced with forwarded via SSH

This is unnecessary and confusing. Future code changes might introduce a
common feature toggle to disable code paths simultaneously in QEMU, HA and
GA setup, but for now it at least will make both code paths consistent.

Signed-off-by: Arthur Sengileyev <arthur.sengileyev@gmail.com>