limactl disk: Do not use qemu-img for raw disks
Fix `limactl disk create` and `limactl disk resize` to use nativeimgutil
for creating and resizing disks using raw format.

There are special cases (using direct I/O on certain file systems) when
creating a raw image should be done with qemu-img, but these are not
relevant to lima on macOS.

With this change we have no dependency on qemu when using VZ instance.

This change does not fix the issue of default "qcow2" disk format.
Users must specify the disk format when creating additional disks for VZ
instance:

    limactl disk create --format raw

Unfinished:
- Needs tests for nativeimgutil and disk command

Fixes: #2853
Signed-off-by: Nir Soffer <nirsof@gmail.com>