Checking for protected dirs must be against mountPoint and not location
The validation code was written at a time when all mounts were mounted at
`location` inside the guest because there was no separate `mountPoint` setting.

Now that `mountPoint` exists, the only validation for `location` is that
the path exists on the host and designates a directory. All validation to
protect guest directories must be done against `mountPoint`.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>