Don't use the result of yaml.Unmarshal after it returns an error
We calll yaml.UnmarshalWithOptions twice, to detect non-strict YAML, and
overwrite the result of the first unmarshal with the second one. This seems
to work, but there is no guarantee that it couldn't have left the struct in
an undefined state.

Also the prototype should use `*LibYAML` instead of `any` because that is the
only valid type you should pass in.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>