disk: Round up disk size to 512 bytes
Apple virtualization framework rejects disks if the disk size is not
aligned to 512 bytes. QEMU block layer round the size up when creating
or resizing disks. Let's do the same to ensure that we don't create or
resize disk to unusable size.

I added only trivial tests fro the RoundUp helper since this kind of
code is easy to get wrong. We need to add more tests for creating and
resizing disk images.

Fixes #3390

Signed-off-by: Nir Soffer <nirsof@gmail.com>