Fix slow image pull caused by delayed DNS response
When IPv6 is disabled, the hostAgent enters a code path that introduces a 1-second sleep whenever
it receives an `AAAA` DNS query. This delay can cause the DNS query to timeout, triggering a retry
loop on the client side. Docker, for example, retries up to 10 times at 1-second intervals—resulting
in a significant delay before an image is pulled.

A suggested workaround is to remove the sleep, which was originally introduced to address a specific
issue that now appears to be outdated or no longer relevant.

[Unexpected A record DNS response](https://github.com/rancher-sandbox/rancher-desktop/issues/3447#issuecomment-1330876897)

Signed-off-by: Nino Kodabande <nkodabande@suse.com>