Locate the limactl executable via os.Args[0] instead of os.Executable()
We need the symlink directory to find the the ../share/lima directory
relative to it. os.Executable() uses /proc/self/exe on Linux and returns
the symlink target, which will break things with a Homebrew installation.

This commit also removes the dead code that pretends to resolve any
symlinks. It never worked because it called os.Stat() instead of os.Lstat().
If it had worked, then code would have been broken on macOS too.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>