Allow dotted instance names in `limactl start`
The code in limatmpl.Read() needs to read non-template files too,
so no longer required a name to have a .yml or .yaml file extension.

This commit changes the `limactl start` and `limactl create` commands
to not call `limatmpl.Read` when the argument looks like a filename,
but doesn't have a .yml or .yaml extension. This preserves backwards
compatibility for commands like `limactl start example.com`.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>