Allow dotted instance names in `limactl start`
limatmpl.Read() no longer returns an empty template when called with
an instance name because it now needs to be able to read any kind of
file via the template:// mechanism.

This commit changes the callers to no longer call limatmpl.Read() with
an instance name. This also allows simplification of the locator type
selection: every locator that is not a URL can now be treated as a local
file.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>