go.mod: drop dependency on github.com/containerd/containerd
Silences govulncheck false positives GO-2025-3699 and GO-2025-3701.

The dependency was also problematic because it looked like as if it
was the actual daemon version running in the VM.

Copied https://github.com/containerd/containerd/tree/v2.1.1/pkg/identifiers
with the modification to drop the dependency on other containerd packages.

The NOTICE file was added for compliance of the section 4(d) of the
Apache License Version 2.0.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>