Search for the data directory using argv[0] *and* os.Executable()
Some version managers like aqua rely on shims in $PATH to run the proper
application. This isn't particularly unusual, but it was broken by the
changes in #3566, because directory lookup ends up pointing to *the
shim* rather than the underlying installation.

The existing logic already has the concept of falling back to
os.Executable if a path based on `argv[0]` can't be determined, so this
issue can be solved by just always checking os.Executable *in addition*
to `argv[0]`.

Signed-off-by: re:fi.64 <hello@refi64.dev>