guestagent: Fix isEmptyEvent check
The intention seems to be checking if an event is equal to empty event,
ignoring the event time, by copying the event and setting its time to
nil. However instead of copying the event, we use the same object and
set the event time to nil.

I don't know if this caused a user visible problem, but it likely log
wrong event time if we log the event.

Signed-off-by: Nir Soffer <nirsof@gmail.com>