Get rid of `fetch-depth: 1`
It is redundant because it is already the default.

I suspect that we used it because in many places we needed to specify
`fetch-depth: 0` to get all the tags. But with one exception for the
colima test, this is no longer the case, so I think we should get rid
of all of these.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>