Add two fields for images exported from containers
The original container image has some optional metadata in the
ENTRYPOINT and STOPSIGNAL fields, that could be needed later...

But these are not exported in the generic rootfs, by default.

Make it possible to pass these at runtime, for instance to
select between OpenRC and systemd - or SIGTERM and systemd.

Some drivers include their config in the rootfs, like wsl.conf.

Signed-off-by: Anders F Björklund <anders.f.bjorklund@gmail.com>