`--progress`: start printing log from the beginning of the current boot session
Fix #3907

The beginning of the current boot session is detected by using /^Cloud-init.* finished/ as an indicator.
It expects that the built-in `awk` exists.

note:
Detecting the current boot session fails on a VM created with template://alpine, because cloud-init-output.log in that VM records some log entry after the recorded indicator.

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>