Add yq subcommands to limactl and lima-guestagent
Adding it to `lima-guestagent` allows us to use `yq` in provisioning
scripts without having to install it from a package repository or
downloading it from GitHub.

Adding it to `limactl` makes it available to integration tests without
having to install it. It is essentially free because we already use
`yqlib` inside `limactl` anyways, , so the executable size doesn't
really change.

The size of the (compressed) guestagent for aarch64 increases from
12MB to 14MB, which seems acceptable.

These are "hidden" commands that don't show up in --help output.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>