Add --yq option to `limactl list` and `limactl info`
The --yq option is only compatible with JSON or YAML output and will
apply the yq expression to the result before writing it to the output.

The following 2 commands are functionally identical (but may result in different
formatting, see below):

    limactl ls default --yq .dir
    limactl ls default --format json | limactl yq .dir

The output of `limactl list --json` will be pretty-printed when it goes
to a terminal. This means it is no longer one line per JSON object, but
has much improved legibility. Output to a file or pipe maintains JSON
Lines rules.

Also colorize the output of `limactl list`, `limactl info`, `limactl tmpl copy`,
and `limactl tmpl yq` if the output goes to the terminal.

Unfortunately the output of `yamlfmt` cannot be colorized, so the terminal
output may look slightly different from the output to a file or pipe (mostly
the extra indentation of list elements), but this seems like a reasonable
compromise.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>