Allow multiple `--yq` options for `limactl ls`
This is mostly useful for selection, so e.g. a plugin can select all
running instances, and an option to the plugin can narrow down the
selection further, or select fields etc.

This change also make the `--quiet` option compatible with `--yq`.
Before `--quiet` took precedence and `--yq` was ignored instead of
generated an incompatibility warning. Now `--quiet` works with `--yq`
by passing the final result through ` | .name`, which implements the
same functionality it normally has.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>